Move xai from chains to testHelper
